requirements.txt
alembic==1.4.2
aniso8601==8.0.0
autopep8==1.5.3
bcrypt==3.1.7
certifi==2020.4.5.2
cffi==1.14.0
click==7.1.2
cloudinary==1.21.1
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.17
SQLAlchemy-serializer==1.3.4.2
toml==0.10.1
urllib3==1.25.9
Werkzeug==1.0.1
